+++ cli/tests/test_backout_command.rs
@@ -38,7 +38,12 @@ fn test_backout() {
 
     // Backout the commit
     let output = work_dir.run_jj(["backout", "-r", "@"]);
-    insta::assert_snapshot!(output, @"");
+    insta::assert_snapshot!(output, @r"
+    ------- stderr -------
+    Warning: `jj backout` is deprecated; use `jj revert` instead
+    Warning: `jj backout` will be removed in a future version, and this will be a hard error
+    [EOF]
+    ");
     insta::assert_snapshot!(get_log_output(&work_dir), @r#"
     ○  6d845ed9fb6a Back out "a"
     │
@@ -56,7 +61,12 @@ fn test_backout() {
     // Backout the new backed-out commit
     work_dir.run_jj(["edit", "@+"]).success();
     let output = work_dir.run_jj(["backout", "-r", "@"]);
-    insta::assert_snapshot!(output, @"");
+    insta::assert_snapshot!(output, @r"
+    ------- stderr -------
+    Warning: `jj backout` is deprecated; use `jj revert` instead
+    Warning: `jj backout` will be removed in a future version, and this will be a hard error
+    [EOF]
+    ");
     insta::assert_snapshot!(get_log_output(&work_dir), @r#"
     ○  79555ea9040b Back out "Back out "a""
     │
@@ -100,7 +110,12 @@ fn test_backout_multiple() {
 
     // Backout multiple commits
     let output = work_dir.run_jj(["backout", "-r", "b", "-r", "c", "-r", "e"]);
-    insta::assert_snapshot!(output, @"");
+    insta::assert_snapshot!(output, @r"
+    ------- stderr -------
+    Warning: `jj backout` is deprecated; use `jj revert` instead
+    Warning: `jj backout` will be removed in a future version, and this will be a hard error
+    [EOF]
+    ");
     insta::assert_snapshot!(get_log_output(&work_dir), @r#"
     ○  6504c4ded177 Back out "b"
     │
@@ -203,7 +218,12 @@ fn test_backout_description_template() {
 
     // Verify that message of backed out commit follows the template
     let output = work_dir.run_jj(["backout", "-r", "a"]);
-    insta::assert_snapshot!(output, @"");
+    insta::assert_snapshot!(output, @r"
+    ------- stderr -------
+    Warning: `jj backout` is deprecated; use `jj revert` instead
+    Warning: `jj backout` will be removed in a future version, and this will be a hard error
+    [EOF]
+    ");
     insta::assert_snapshot!(get_log_output(&work_dir), @r#"
     ○  1db880a5204e Revert commit 2443ea76b0b1 "a"
     @  2443ea76b0b1 a